refactor: unify upgrade with installation and refactor router into individual helpers
