--- packages/aio-commerce-lib-app/source/actions/installation/schema.ts
@@ -1,27 +1,0 @@
-/*
- * Copyright 2026 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-import * as v from "valibot";
-
-import { LifecycleRequestContextSchema } from "#management/common/schema";
-
-/** Request body for POST / and POST /validation — the shared lifecycle request shape. */
-export const InstallationRequestBodySchema = LifecycleRequestContextSchema;
-
-const {
-  commerceBaseUrl: _,
-  commerceEnv: __,
-  ...upgradeSchema
-} = LifecycleRequestContextSchema.entries;
-
-/** Request body for POST /upgrade. */
-export const UpgradeRequestBodySchema = v.object(upgradeSchema);
